+++ src/test/converter2/behavior/cascadedModifiers.ts
@@ -1,0 +1,14 @@
+/**
+ * @beta
+ */
+export namespace BetaStuff {
+    export class AlsoBeta {
+        betaFish() {}
+
+        /** @alpha */
+        alphaFish() {}
+    }
+}
+
+/** @alpha @beta */
+export const mutuallyExclusive = true;